# Interface cast spin lock