--- a/src/bootstrap/gerbil/compiler/base__0.scm
+++ b/src/bootstrap/gerbil/compiler/base__0.scm
@@ -1,6 +1,6 @@
 (declare (block) (standard-bindings) (extended-bindings))
 (begin
-  (define gerbil/compiler/base::timestamp 1708387688)
+  (define gerbil/compiler/base::timestamp 1708418155)
   (begin
     (define gxc#current-compile-symbol-table (make-parameter '#f))
     (define gxc#current-compile-runtime-sections (make-parameter '#f))
@@ -18,25 +18,25 @@
     (define gxc#current-compile-decls (make-parameter '#f))
     (define gxc#current-compile-context (make-parameter '#f))
     (define gxc#symbol-table::t
-      (let ((__tmp97180 (list))
-            (__tmp97178
-             (let ((__tmp97179
+      (let ((__tmp97249 (list))
+            (__tmp97247
+             (let ((__tmp97248
                     (let () (declare (not safe)) (cons 'struct: '#t))))
                (declare (not safe))
-               (cons __tmp97179 '()))))
+               (cons __tmp97248 '()))))
         (declare (not safe))
         (make-class-type
          'gxc#symbol-table::t
          'symbol-table
-         __tmp97180
+         __tmp97249
          '(gensyms bindings)
-         __tmp97178
+         __tmp97247
          ':init!)))
     (define gxc#symbol-table?
       (let () (declare (not safe)) (make-class-predicate gxc#symbol-table::t)))
     (define gxc#make-symbol-table
-      (lambda _$args97131_
-        (apply make-instance gxc#symbol-table::t _$args97131_)))
+      (lambda _$args97200_
+        (apply make-instance gxc#symbol-table::t _$args97200_)))
     (define gxc#symbol-table-gensyms
       (let ()
         (declare (not safe))
@@ -70,49 +70,49 @@
         (declare (not safe))
         (make-class-slot-unchecked-mutator gxc#symbol-table::t 'bindings)))
     (define gxc#symbol-table:::init!
-      (lambda (_self97129_)
-        (if (let ((__tmp97186
+      (lambda (_self97198_)
+        (if (let ((__tmp97255
                    (let ()
                      (declare (not safe))
-                     (##structure-length _self97129_))))
+                     (##structure-length _self97198_))))
               (declare (not safe))
-              (##fx< '2 __tmp97186))
+              (##fx< '2 __tmp97255))
             (begin
-              (let ((__tmp97183
+              (let ((__tmp97252
                      (let () (declare (not safe)) (make-hash-table-eq)))
-                    (__tmp97182
+                    (__tmp97251
                      (let ()
                        (declare (not safe))
-                       (##structure-type _self97129_))))
+                       (##structure-type _self97198_))))
                 (declare (not safe))
                 (##unchecked-structure-set!
-                 _self97129_
-                 __tmp97183
+                 _self97198_
+                 __tmp97252
                  '1
-                 __tmp97182
+                 __tmp97251
                  '#f))
-              (let ((__tmp97185
+              (let ((__tmp97254
                      (let () (declare (not safe)) (make-hash-table-eq)))
-                    (__tmp97184
+                    (__tmp97253
                      (let ()
                        (declare (not safe))
-                       (##structure-type _self97129_))))
+                       (##structure-type _self97198_))))
                 (declare (not safe))
                 (##unchecked-structure-set!
-                 _self97129_
-                 __tmp97185
+                 _self97198_
+                 __tmp97254
                  '2
-                 __tmp97184
+                 __tmp97253
                  '#f)))
-            (let ((__tmp97181
+            (let ((__tmp97250
                    (let ()
                      (declare (not safe))
-                     (##vector-length _self97129_))))
+                     (##vector-length _self97198_))))
               (declare (not safe))
               (error '"struct-instance-init!: too many arguments for struct"
-                     _self97129_
+                     _self97198_
                      '2
-                     __tmp97181)))))
+                     __tmp97250)))))
     (let ()
       (declare (not safe))
       (bind-method!__%
@@ -121,46 +121,46 @@
        gxc#symbol-table:::init!
        '#f))
     (define gxc#raise-compile-error
-      (lambda (_message96998_ _stx96999_ . _details97000_)
-        (let ((_ctx97005_
-               (let ((_$e97002_ (gxc#current-compile-context)))
-                 (if _$e97002_ _$e97002_ 'compile))))
+      (lambda (_message97067_ _stx97068_ . _details97069_)
+        (let ((_ctx97074_
+               (let ((_$e97071_ (gxc#current-compile-context)))
+                 (if _$e97071_ _$e97071_ 'compile))))
           (apply gx#raise-syntax-error
-                 _ctx97005_
-                 _message96998_
-                 _stx96999_
-                 _details97000_))))
+                 _ctx97074_
+                 _message97067_
+                 _stx97068_
+                 _details97069_))))
     (define gxc#verbose
-      (lambda _args96995_
+      (lambda _args97064_
         (if (gxc#current-compile-verbose)
-            (let ((__tmp97187 (lambda () (apply displayln _args96995_))))
+            (let ((__tmp97256 (lambda () (apply displayln _args97064_))))
               (declare (not safe))
-              (with-lock gxc#+verbose-mutex+ __tmp97187))
+              (with-lock gxc#+verbose-mutex+ __tmp97256))
             '#!void)))
     (define gxc#+verbose-mutex+ (make-mutex 'compiler/driver))
     (define gxc#module-path-reserved-chars '":#<>&!?*;()[]{}|'`\"\\")
     (define gxc#module-id->path-string
-      (lambda (_id96977_)
-        (let* ((_str96979_
-                (if (let () (declare (not safe)) (symbol? _id96977_))
-                    (symbol->string _id96977_)
-                    _id96977_))
-               (_len96981_ (string-length _str96979_))
-               (_res96983_ (make-string _len96981_)))
-          (let _lp96986_ ((_i96988_ '0))
-            (if (fx< _i96988_ _len96981_)
-                (let* ((_char96990_ (string-ref _str96979_ _i96988_))
-                       (_xchar96992_
+      (lambda (_id97046_)
+        (let* ((_str97048_
+                (if (let () (declare (not safe)) (symbol? _id97046_))
+                    (symbol->string _id97046_)
+                    _id97046_))
+               (_len97050_ (string-length _str97048_))
+               (_res97052_ (make-string _len97050_)))
+          (let _lp97055_ ((_i97057_ '0))
+            (if (fx< _i97057_ _len97050_)
+                (let* ((_char97059_ (string-ref _str97048_ _i97057_))
+                       (_xchar97061_
                         (if (let ()
                               (declare (not safe))
                               (string-index__0
                                gxc#module-path-reserved-chars
-                               _char96990_))
+                               _char97059_))
                             '#\_
-                            _char96990_)))
-                  (string-set! _res96983_ _i96988_ _xchar96992_)
-                  (let ((__tmp97188
-                         (let () (declare (not safe)) (fx+ _i96988_ '1))))
+                            _char97059_)))
+                  (string-set! _res97052_ _i97057_ _xchar97061_)
+                  (let ((__tmp97257
+                         (let () (declare (not safe)) (fx+ _i97057_ '1))))
                     (declare (not safe))
-                    (_lp96986_ __tmp97188)))
-                _res96983_)))))))
+                    (_lp97055_ __tmp97257)))
+                _res97052_)))))))