--- a/src/bootstrap/gerbil/compiler/driver__1.scm
+++ b/src/bootstrap/gerbil/compiler/driver__1.scm
@@ -1,169 +1,169 @@
 (declare (block) (standard-bindings) (extended-bindings) (inlining-limit 200))
 (begin
   (define |gxc[:0:]#with-driver-mutex|
-    (lambda (_$stx156164_)
-      (let* ((_g156168156182_
-              (lambda (_g156169156178_)
+    (lambda (_$stx156233_)
+      (let* ((_g156237156251_
+              (lambda (_g156238156247_)
                 (let ()
                   (declare (not safe))
                   (gx#raise-syntax-error
                    '#f
                    '"Bad syntax; invalid match target"
-                   _g156169156178_))))
-             (_g156167156224_
-              (lambda (_g156169156186_)
+                   _g156238156247_))))
+             (_g156236156293_
+              (lambda (_g156238156255_)
                 (if (let ()
                       (declare (not safe))
-                      (gx#stx-pair? _g156169156186_))
-                    (let ((_e156173156189_
+                      (gx#stx-pair? _g156238156255_))
+                    (let ((_e156242156258_
                            (let ()
                              (declare (not safe))
-                             (gx#syntax-e _g156169156186_))))
-                      (let ((_hd156172156193_
+                             (gx#syntax-e _g156238156255_))))
+                      (let ((_hd156241156262_
                              (let ()
                                (declare (not safe))
-                               (##car _e156173156189_)))
-                            (_tl156171156196_
+                               (##car _e156242156258_)))
+                            (_tl156240156265_
                              (let ()
                                (declare (not safe))
-                               (##cdr _e156173156189_))))
+                               (##cdr _e156242156258_))))
                         (if (let ()
                               (declare (not safe))
-                              (gx#stx-pair? _tl156171156196_))
-                            (let ((_e156176156199_
+                              (gx#stx-pair? _tl156240156265_))
+                            (let ((_e156245156268_
                                    (let ()
                                      (declare (not safe))
-                                     (gx#syntax-e _tl156171156196_))))
-                              (let ((_hd156175156203_
+                                     (gx#syntax-e _tl156240156265_))))
+                              (let ((_hd156244156272_
                                      (let ()
                                        (declare (not safe))
-                                       (##car _e156176156199_)))
-                                    (_tl156174156206_
+                                       (##car _e156245156268_)))
+                                    (_tl156243156275_
                                      (let ()
                                        (declare (not safe))
-                                       (##cdr _e156176156199_))))
+                                       (##cdr _e156245156268_))))
                                 (if (let ()
                                       (declare (not safe))
-                                      (gx#stx-null? _tl156174156206_))
-                                    ((lambda (_L156209_)
-                                       (let ((__tmp157965
+                                      (gx#stx-null? _tl156243156275_))
+                                    ((lambda (_L156278_)
+                                       (let ((__tmp158034
                                               (let ()
                                                 (declare (not safe))
                                                 (gx#datum->syntax__0
                                                  '#f
                                                  'with-lock)))
-                                             (__tmp157958
-                                              (let ((__tmp157964
+                                             (__tmp158027
+                                              (let ((__tmp158033
                                                      (let ()
                                                        (declare (not safe))
                                                        (gx#datum->syntax__0
                                                         '#f
                                                         '+driver-mutex+)))
-                                                    (__tmp157959
-                                                     (let ((__tmp157960
-                                                            (let ((__tmp157963
+                                                    (__tmp158028
+                                                     (let ((__tmp158029
+                                                            (let ((__tmp158032
 ;;<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
                            (let ()
                              (declare (not safe))
                              (gx#datum->syntax__0 '#f 'lambda)))
-                          (__tmp157961
-                           (let ((__tmp157962
+                          (__tmp158030
+                           (let ((__tmp158031
                                   (let ()
                                     (declare (not safe))
-                                    (cons _L156209_ '()))))
+                                    (cons _L156278_ '()))))
                              (declare (not safe))
-                             (cons '() __tmp157962))))
+                             (cons '() __tmp158031))))
                       (declare (not safe))
-                      (cons __tmp157963 __tmp157961))))
+                      (cons __tmp158032 __tmp158030))))
                (declare (not safe))
-               (cons __tmp157960 '()))))
+               (cons __tmp158029 '()))))
 ;;>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
                                                 (declare (not safe))
-                                                (cons __tmp157964
-                                                      __tmp157959))))
+                                                (cons __tmp158033
+                                                      __tmp158028))))
                                          (declare (not safe))
-                                         (cons __tmp157965 __tmp157958)))
-                                     _hd156175156203_)
-                                    (_g156168156182_ _g156169156186_))))
-                            (_g156168156182_ _g156169156186_))))
-                    (_g156168156182_ _g156169156186_)))))
-        (_g156167156224_ _$stx156164_))))
+                                         (cons __tmp158034 __tmp158027)))
+                                     _hd156244156272_)
+                                    (_g156237156251_ _g156238156255_))))
+                            (_g156237156251_ _g156238156255_))))
+                    (_g156237156251_ _g156238156255_)))))
+        (_g156236156293_ _$stx156233_))))
   (define |gxc[:0:]#go!|
-    (lambda (_$stx156228_)
-      (let* ((_g156232156246_
-              (lambda (_g156233156242_)
+    (lambda (_$stx156297_)
+      (let* ((_g156301156315_
+              (lambda (_g156302156311_)
                 (let ()
                   (declare (not safe))
                   (gx#raise-syntax-error
                    '#f
                    '"Bad syntax; invalid match target"
-                   _g156233156242_))))
-             (_g156231156287_
-              (lambda (_g156233156250_)
+                   _g156302156311_))))
+             (_g156300156356_
+              (lambda (_g156302156319_)
                 (if (let ()
                       (declare (not safe))
-                      (gx#stx-pair? _g156233156250_))
-                    (let ((_e156237156253_
+                      (gx#stx-pair? _g156302156319_))
+                    (let ((_e156306156322_
                            (let ()
                              (declare (not safe))
-                             (gx#syntax-e _g156233156250_))))
-                      (let ((_hd156236156257_
+                             (gx#syntax-e _g156302156319_))))
+                      (let ((_hd156305156326_
                              (let ()
                                (declare (not safe))
-                               (##car _e156237156253_)))
-                            (_tl156235156260_
+                               (##car _e156306156322_)))
+                            (_tl156304156329_
                              (let ()
                                (declare (not safe))
-                               (##cdr _e156237156253_))))
+                               (##cdr _e156306156322_))))
                         (if (let ()
                               (declare (not safe))
-                              (gx#stx-pair? _tl156235156260_))
-                            (let ((_e156240156263_
+                              (gx#stx-pair? _tl156304156329_))
+                            (let ((_e156309156332_
                                    (let ()
                                      (declare (not safe))
-                                     (gx#syntax-e _tl156235156260_))))
-                              (let ((_hd156239156267_
+                                     (gx#syntax-e _tl156304156329_))))
+                              (let ((_hd156308156336_
                                      (let ()
                                        (declare (not safe))
-                                       (##car _e156240156263_)))
-                                    (_tl156238156270_
+                                       (##car _e156309156332_)))
+                                    (_tl156307156339_
                                      (let ()
                                        (declare (not safe))
-                                       (##cdr _e156240156263_))))
+                                       (##cdr _e156309156332_))))
                                 (if (let ()
                                       (declare (not safe))
-                                      (gx#stx-null? _tl156238156270_))
-                                    ((lambda (_L156273_)
-                                       (let ((__tmp157971
+                                      (gx#stx-null? _tl156307156339_))
+                                    ((lambda (_L156342_)
+                                       (let ((__tmp158040
                                               (let ()
                                                 (declare (not safe))
                                                 (gx#datum->syntax__0
                                                  '#f
                                                  'spawn)))
-                                             (__tmp157966
-                                              (let ((__tmp157967
-                                                     (let ((__tmp157970
+                                             (__tmp158035
+                                              (let ((__tmp158036
+                                                     (let ((__tmp158039
                                                             (let ()
                                                               (declare
                                                                 (not safe))
                                                               (gx#datum->syntax__0
                                                                '#f
                                                                'lambda)))
-                                                           (__tmp157968
-                                                            (let ((__tmp157969
+                                                           (__tmp158037
+                                                            (let ((__tmp158038
 ;;<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
-                           (let () (declare (not safe)) (cons _L156273_ '()))))
+                           (let () (declare (not safe)) (cons _L156342_ '()))))
                       (declare (not safe))
-                      (cons '() __tmp157969))))
+                      (cons '() __tmp158038))))
                (declare (not safe))
-               (cons __tmp157970 __tmp157968))))
+               (cons __tmp158039 __tmp158037))))
 ;;>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
                                                 (declare (not safe))
-                                                (cons __tmp157967 '()))))
+                                                (cons __tmp158036 '()))))
                                          (declare (not safe))
-                                         (cons __tmp157971 __tmp157966)))
-                                     _hd156239156267_)
-                                    (_g156232156246_ _g156233156250_))))
-                            (_g156232156246_ _g156233156250_))))
-                    (_g156232156246_ _g156233156250_)))))
-        (_g156231156287_ _$stx156228_)))))
+                                         (cons __tmp158040 __tmp158035)))
+                                     _hd156308156336_)
+                                    (_g156301156315_ _g156302156319_))))
+                            (_g156301156315_ _g156302156319_))))
+                    (_g156301156315_ _g156302156319_)))))
+        (_g156300156356_ _$stx156297_)))))