--- a/src/bootstrap/gerbil/core$_MOP_$_MOP_2___0.scm
+++ b/src/bootstrap/gerbil/core$_MOP_$_MOP_2___0.scm
@@ -1,6 +1,6 @@
 (declare (block) (standard-bindings) (extended-bindings))
 (begin
-  (define gerbil/core$<MOP>$<MOP:2>::timestamp 1708387685)
+  (define gerbil/core$<MOP>$<MOP:2>::timestamp 1708418151)
   (begin
     (define gerbil/core$<MOP>$<MOP:2>#class-type-info::t
       (let ((__tmp49651 (list)))