--- a/src/bootstrap/gerbil/core__14.scm
+++ b/src/bootstrap/gerbil/core__14.scm
@@ -21,23 +21,23 @@
                                (declare (not safe))
                                (##cdr _e3455934572_))))
                         ((lambda (_L34582_)
-                           (let ((__tmp52931
+                           (let ((__tmp52908
                                   (gx#datum->syntax '#f 'make-setq-macro))
-                                 (__tmp52927
-                                  (let ((__tmp52928
-                                         (let ((__tmp52929
-                                                (let ((__tmp52930
+                                 (__tmp52904
+                                  (let ((__tmp52905
+                                         (let ((__tmp52906
+                                                (let ((__tmp52907
                                                        (gx#datum->syntax
                                                         '#f
                                                         'syntax-rules)))
                                                   (declare (not safe))
-                                                  (cons __tmp52930 _L34582_))))
+                                                  (cons __tmp52907 _L34582_))))
                                            (declare (not safe))
-                                           (cons __tmp52929 '()))))
+                                           (cons __tmp52906 '()))))
                                     (declare (not safe))
-                                    (cons 'macro: __tmp52928))))
+                                    (cons 'macro: __tmp52905))))
                              (declare (not safe))
-                             (cons __tmp52931 __tmp52927)))
+                             (cons __tmp52908 __tmp52904)))
                          _tl3455734579_)))
                     (_g3455434565_ _g3455534569_)))))
         (_g3455334595_ _$stx34550_))))