--- a/src/bootstrap/gerbil/core__16.scm
+++ b/src/bootstrap/gerbil/core__16.scm
@@ -32,10 +32,10 @@
                                      (if (gx#current-expander-compiling?)
                                          (gx#eval-syntax _L37262_)
                                          '#!void)
-                                     (let ((__tmp53033
+                                     (let ((__tmp53010
                                             (gx#datum->syntax '#f 'void)))
                                        (declare (not safe))
-                                       (cons __tmp53033 '())))
+                                       (cons __tmp53010 '())))
                                    _hd3722837256_)
                                   (_g3722137235_ _g3722237239_))))
                           (_g3722137235_ _g3722237239_))))