--- a/src/bootstrap/gerbil/expander/common__0.scm
+++ b/src/bootstrap/gerbil/expander/common__0.scm
@@ -1,26 +1,26 @@
 (declare (block) (standard-bindings) (extended-bindings))
 (begin
-  (define gerbil/expander/common::timestamp 1708387688)
+  (define gerbil/expander/common::timestamp 1708418154)
   (begin
     (define gx#AST::t
-      (let ((__tmp79055 (list))
-            (__tmp79053
-             (let ((__tmp79054
+      (let ((__tmp79124 (list))
+            (__tmp79122
+             (let ((__tmp79123
                     (let () (declare (not safe)) (cons 'struct: '#t))))
                (declare (not safe))
-               (cons __tmp79054 '()))))
+               (cons __tmp79123 '()))))
         (declare (not safe))
         (make-class-type
          'gerbil#AST::t
          'syntax
-         __tmp79055
+         __tmp79124
          '(e source)
-         __tmp79053
+         __tmp79122
          '#f)))
     (define gx#AST?
       (let () (declare (not safe)) (make-class-predicate gx#AST::t)))
     (define gx#make-AST
-      (lambda _$args78952_ (apply make-instance gx#AST::t _$args78952_)))
+      (lambda _$args79021_ (apply make-instance gx#AST::t _$args79021_)))
     (define gx#AST-e
       (let () (declare (not safe)) (make-class-slot-accessor gx#AST::t 'e)))
     (define gx#AST-source