--- a/src/bootstrap/gerbil/runtime/gambit__0.scm
+++ b/src/bootstrap/gerbil/runtime/gambit__0.scm
@@ -1,2 +1,2 @@
 (declare (block) (standard-bindings) (extended-bindings))
-(begin (define gerbil/runtime/gambit::timestamp 1708387687) '#!void)
+(begin (define gerbil/runtime/gambit::timestamp 1708418153) '#!void)