--- a/src/bootstrap/gerbil/runtime/table__rt.scm
+++ b/src/bootstrap/gerbil/runtime/table__rt.scm
@@ -1,4 +1,6 @@
 (declare (block) (standard-bindings) (extended-bindings))
 (begin
-  (load-module "gerbil/runtime/gambit__rt")
+  (begin
+    (load-module "gerbil/runtime/gambit__rt")
+    (load-module "gerbil/runtime/util__rt"))
   (load-module "gerbil/runtime/table__0"))