--- a/src/bootstrap/gerbil/runtime__0.scm
+++ b/src/bootstrap/gerbil/runtime__0.scm
@@ -1,2 +1,2 @@
 (declare (block) (standard-bindings) (extended-bindings))
-(begin (define gerbil/runtime::timestamp 1708387688) '#!void)
+(begin (define gerbil/runtime::timestamp 1708418154) '#!void)