This changes the prototype table lock to be a spin lock .... and it is almost 3x faster.

Before:
```
$ /tmp/interface-bench cast 10 1000000
(time (let () (declare (not safe)) (std/interface-benchmark#cast-benchmark _iters180_ _threads179_ std/interface-benchmark#do-cast)))
    2.117610 secs real time
    2.117559 secs cpu time (2.117150 user, 0.000409 system)
    142 collections accounting for 0.122705 secs real time (0.122418 user, 0.000250 system)
    1120026112 bytes allocated
    672 minor faults
    no major faults
    5529541900 cpu cycles
```

After:
```
$ /tmp/interface-bench2 cast 10 1000000
(time (let () (declare (not safe)) (std/interface-benchmark#cast-benchmark _iters180_ _threads179_ std/interface-benchmark#do-cast)))
    0.857509 secs real time
    0.857492 secs cpu time (0.857288 user, 0.000204 system)
    142 collections accounting for 0.104038 secs real time (0.103984 user, 0.000000 system)
    1119990592 bytes allocated
    672 minor faults
    no major faults
    2239135638 cpu cycles
```