@fare 2024-02-20
I think we should be using futexes or something else on Linux that doesn't busywait and waste batteries on laptop smp—and importantly, this should be all abstracted over somehow in a macro.

If you want this in, fine for now, but then you should open an issue about getting locking right on SMP.