@vyzo 2024-02-20
It is abstracted over a macro, see `__lock-inline!` and `__unlock-inline!`.

It is a workable solution for SMP, the critical sections are expected to be small (less than 10ns each) so the spin/busy wait is not that bad.

Ideally we'd have futexes; unfortunately full gambit mutexes are so much slower it is not even funny.